\name{YeastmRNACor-package}
\alias{YeastmRNACor-package}
\alias{YeastmRNACor}
\docType{package}
\title{
Yeast Correlation Finder
}
\description{
Find S.cerevisiae genes with correlated expression
}
\details{
\tabular{ll}{
  Package: \tab YeastmRNACor\cr
  Type: \tab Package\cr
  Version: \tab 0.99.0\cr
  Date: \tab 2012-10-12\cr
  License: \tab Artistic-2.0\cr
  }
}
\author{
Paul Shannon
Maintainer: Paul Shannon <pshannon@fhcrc.org>
}
\references{
 Allocco et al, 2004, "Quantifying the relationship between
 co-expression, co-regulation and gene function":
}
\keyword{manip}
